Disable default plotting of regular's subsampling
Since the main focus on the paper is on analysing the adjustable
functions, no subsampling figures are included in the paper and
most possible files are not generated either.

This means that for these plot scripts to make sense, the associated
cases should be enabled in the experiment scripts and run.