Plot_error_grid title_width now as argument
Different line-wrap widths are needed for different plots, but that
depends on the figure size, presence of y-axis ticks/labels and
colorbar, and the length of tht title.

So by making the desired width a function argument, the finetuning
can be moved to the exact script function that defines the other
plot arguments too.