Limit GitHub actions testing to python 3.9 only
Main reason for this decision are type hints. Prior to 3.9, generic
types such as lists and dictionaries had to be typed using typing.List
or typing.Dict, but builtin 'list' and 'dict' can be used since 3.9.

Since this is 'simply' application code and not a library, there is no
additional benefit to testing for 3.6, 3.7 or 3.8 if experiments are
only going to be run with 3.9 (or maybe 3.10) anyway